restructure component folder and separate styles
