+++ src/components/checkbox/styles.tsx
@@ -1,0 +1,14 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flexDirection: "row",
+        alignItems: "center"
+    },
+    boxBase: {
+        alignItems: "center",
+        justifyContent: "center",
+        overflow: "hidden",
+    },
+});
+export default styles;
