+++ src/components/checkbox/types.ts
@@ -6,13 +6,13 @@ import type {
     RegisteredStyle,
     ViewStyle
 } from "react-native";
-import type React from "react";
+import type {ReactNode} from "react";
 
 export type CheckboxProps = {
     onPress?: ((_: GestureResponderEvent) => void);
     style?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<Falsy | ViewStyle | RegisteredStyle<ViewStyle>> | readonly (Falsy | ViewStyle | RegisteredStyle<ViewStyle>)[];
     boxStyle?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<Falsy | ViewStyle | RegisteredStyle<ViewStyle>> | readonly (Falsy | ViewStyle | RegisteredStyle<ViewStyle>)[];
-    children?: React.ReactChild | React.ReactFragment;
+    children?: ReactNode;
     fill?: boolean;
     backgroundColor?: string;
     borderSize?: number;
